requirements.txt
python-dotenv==1.0.1
flask==3.1.0
llama-index==0.12.22
llama-index-llms-gemini==0.4.11
llama-index-embeddings-gemini==0.3.1
langchain_google_genai==2.0.11
sentence-transformers==3.4.1
gpt-researcher==0.12.8
pydantic